+++ resources/js/lang/en/app.js
@@ -59,7 +59,8 @@ export default {
     rooms: {
       title: 'Rooms',
       create: 'Create rooms',
-      viewAll: 'Show all rooms'
+      viewAll: 'Show all rooms',
+      manage: 'Manage all rooms'
     },
 
     meetings: {