+++ resources/js/lang/en/rooms.js
@@ -93,8 +93,8 @@ export default {
     roles: {
       guest: 'Guest',
       participant: 'Participant',
-      moderator: 'Moderator'
-
+      moderator: 'Moderator',
+      co_owner: 'Co-owner'
     },
     modals: {
       edit: {